Improve readme checks
