+++ parser/tests/check_readme.rs
@@ -1,28 +1,32 @@
-//! Tests that the README code samples actually work.
+//! Tests that the README code sample actually parses.
 
 use pulldown_cmark::{CodeBlockKind, Event, Parser, Tag};
-use rand::{thread_rng, Rng};
 
-use arithmetic_eval::{fns, Assertions, Environment, Prelude, Value, VariableMap};
-use arithmetic_parser::grammars::{F32Grammar, Parse, Untyped};
+use std::fs;
+
+use arithmetic_parser::grammars::{F64Grammar, MockTypes, Parse, WithMockedTypes};
+
+struct MockedTypesList;
+
+impl MockTypes for MockedTypesList {
+    const MOCKED_TYPES: &'static [&'static str] = &["Num"];
+}
+
+type Grammar = WithMockedTypes<F64Grammar, MockedTypesList>;
+
+fn read_file(path: &str) -> String {
+    fs::read_to_string(path).unwrap_or_else(|err| panic!("Cannot read file {}: {}", path, err))
+}
 
 fn check_sample(code_sample: &str) {
-    let program = Untyped::<F32Grammar>::parse_statements(code_sample).unwrap();
-
-    let mut env: Environment<'_, f32> = Prelude.iter().chain(Assertions.iter()).collect();
-    env.insert("array", Value::native_fn(fns::Array)).insert(
-        "rand_num",
-        Value::wrapped_fn(|min: f32, max: f32| thread_rng().gen_range(min..max)),
-    );
-    let module = env.compile_module("test", &program).unwrap();
-    module.run().unwrap();
+    Grammar::parse_statements(code_sample).unwrap();
 }
 
 #[test]
-fn code_samples_in_readme_are_valid() {
-    const README: &str = include_str!("../README.md");
+fn code_sample_in_readme_is_parsed() {
+    let readme = read_file("README.md");
 
-    let parser = Parser::new(README);
+    let parser = Parser::new(&readme);
     let mut code: Option<String> = None;
     for event in parser {
         match event {
